fix(cli): avoid initialization for version queries
